Misc fixes for polys module

+++ diofant/core/evalf.py
@@ -1205,7 +1205,7 @@ def evalf(x, prec, options):
             # convert (approximately) from given tolerance;
             # the formula here will will make 1e-i rounds to 0 for
             # i in the range +/-27 while 2e-i will not be chopped
-            chop_prec = int(round(-3.321*math.log10(chop) + 2.5))
+            chop_prec = round(-3.321*math.log10(chop) + 2.5)
         r = chop_parts(r, chop_prec)
     if options.get("strict"):
         check_target(x, r, prec)
